+++ iscsi-scst/kernel/config.c
@@ -50,8 +50,8 @@ static int iscsi_log_info_show(struct seq_file *seq, void *v)
 	return res;
 }
 
-static int iscsi_proc_log_entry_write(struct file *file, const char __user *buf,
-					 size_t length, loff_t *off)
+static ssize_t iscsi_proc_log_entry_write(struct file *file,
+	const char __user *buf, size_t length, loff_t *off)
 {
 	int res = 0;
 
@@ -475,10 +475,10 @@ void iscsi_dump_iov(struct msghdr *msg)
 {
 	if (trace_flag & TRACE_D_IOV) {
 		int i;
-		printk("%p, %d\n", msg->msg_iov, msg->msg_iovlen);
+		printk("%p, %zd\n", msg->msg_iov, msg->msg_iovlen);
 		for (i = 0; i < min_t(size_t, msg->msg_iovlen, 
 				ISCSI_CONN_IOV_MAX); i++) {
-			printk("%d: %p,%d\n", i, msg->msg_iov[i].iov_base,
+			printk("%d: %p,%zd\n", i, msg->msg_iov[i].iov_base,
 				msg->msg_iov[i].iov_len);
 		}
 	}
@@ -520,7 +520,7 @@ void iscsi_dump_pdu(struct iscsi_pdu *pdu)
 		int i;
 
 		buf = (void *)&pdu->bhs;
-		printk("BHS: (%p,%d)\n", buf, sizeof(pdu->bhs));
+		printk("BHS: (%p,%zd)\n", buf, sizeof(pdu->bhs));
 		for (i = 0; i < sizeof(pdu->bhs); i++)
 			iscsi_dump_char(*buf++);
 		iscsi_dump_char(-1);
